# frontend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WordMind API",
    "description": "单词记忆与思维导图学习平台 API 文档",
    "version": "1.0.0",
    "contact": {
      "name": "WordMind Team",
      "email": "support@wordmind.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/words": {
      "get": {
        "tags": ["Words"],
        "summary": "获取单词列表",
        "description": "分页查询单词列表，支持关键词搜索和词性过滤",
        "operationId": "getWords",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "搜索关键词（单词或释义）",
            "schema": { "type": "string", "example": "ephem" }
          },
          {
            "name": "pos",
            "in": "query",
            "description": "词性过滤",
            "schema": { "type": "string", "example": "adj." }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": { "type": "integer", "default": 1, "example": 1 }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页大小",
            "schema": { "type": "integer", "default": 20, "example": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWordListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/words/{id}": {
      "get": {
        "tags": ["Words"],
        "summary": "获取单词详情",
        "description": "根据 ID 获取单词详情",
        "operationId": "getWordById",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "单词 ID",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWordResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/stats/me": {
      "get": {
        "tags": ["Stats"],
        "summary": "获取学习统计",
        "description": "获取当前用户的学习统计数据",
        "operationId": "getMyStats",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStatsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/study-plans": {
      "post": {
        "tags": ["Stats"],
        "summary": "创建学习计划",
        "description": "为单词创建学习计划",
        "operationId": "createStudyPlan",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudyPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudyPlanResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/study-records": {
      "get": {
        "tags": ["Stats"],
        "summary": "获取学习记录",
        "description": "获取当前用户的学习计划列表",
        "operationId": "getStudyRecords",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListStudyPlanResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/words/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "更新单词",
        "description": "管理员更新单词信息",
        "operationId": "updateWord",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "单词 ID",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWordResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "删除单词",
        "description": "管理员删除单词",
        "operationId": "deleteWord",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "单词 ID",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 认证：在 Authorization header 中填入 Bearer {token}"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "未授权 - 需要有效的 JWT Token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponseVoid"
            },
            "example": {
              "code": 401,
              "message": "未提供有效的认证令牌",
              "data": null,
              "traceId": "abc123xyz"
            }
          }
        }
      },
      "Forbidden": {
        "description": "禁止访问 - 权限不足",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponseVoid"
            },
            "example": {
              "code": 403,
              "message": "没有权限执行此操作",
              "data": null,
              "traceId": "abc123xyz"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponseVoid"
            },
            "example": {
              "code": 404,
              "message": "请求的资源不存在",
              "data": null,
              "traceId": "abc123xyz"
            }
          }
        }
      },
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponseVoid"
            },
            "example": {
              "code": 400,
              "message": "请求参数验证失败",
              "data": null,
              "traceId": "abc123xyz"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "UpdateWordRequest": {
        "type": "object",
        "properties": {
          "word": { "type": "string", "example": "ephemeral", "maxLength": 100 },
          "phonetic": { "type": "string", "example": "/ɪˈfemərəl/", "maxLength": 100 },
          "pos": { "type": "string", "example": "adj.", "maxLength": 50 },
          "meaning": { "type": "string", "example": "短暂的，瞬息的", "maxLength": 500 },
          "example": { "type": "string", "example": "The ephemeral beauty of cherry blossoms.", "maxLength": 1000 },
          "memoryTip": { "type": "string", "example": "e(出)+phem(出现)+eral", "maxLength": 500 }
        }
      },
      "WordResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "word": { "type": "string", "example": "ephemeral" },
          "phonetic": { "type": "string", "example": "/ɪˈfemərəl/" },
          "pos": { "type": "string", "example": "adj." },
          "meaning": { "type": "string", "example": "短暂的，瞬息的" },
          "example": { "type": "string", "example": "The ephemeral beauty of cherry blossoms." },
          "memoryTip": { "type": "string", "example": "e(出)+phem(出现)+eral" },
          "createdAt": { "type": "string", "example": "2024-01-15T10:30:00", "format": "date-time" }
        },
        "required": ["id", "word", "meaning"]
      },
      "WordListResponse": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/WordResponse" }
          },
          "total": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "size": { "type": "integer", "example": 20 }
        },
        "required": ["list", "total", "page", "size"]
      },
...
      "StatsResponse": {
        "type": "object",
        "properties": {
          "totalWords": { "type": "integer", "example": 150 },
          "todayReviewCount": { "type": "integer", "example": 15 },
          "accuracy": { "type": "number", "example": 0.85, "format": "double" },
          "streakDays": { "type": "integer", "example": 7 }
        },
        "required": ["totalWords", "todayReviewCount", "accuracy", "streakDays"]
      },
      "CreateStudyPlanRequest": {
        "type": "object",
        "properties": {
          "wordId": { "type": "integer", "example": 1 },
          "planType": { "type": "string", "example": "vocabulary", "enum": ["vocabulary", "review", "exam"] }
        },
        "required": ["wordId", "planType"]
      },
      "StudyPlanResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "wordId": { "type": "integer", "example": 1 },
          "word": { "type": "string", "example": "ephemeral" },
          "meaning": { "type": "string", "example": "短暂的，瞬息的" },
          "planType": { "type": "string", "example": "vocabulary" },
          "createdAt": { "type": "string", "example": "2024-01-15T10:30:00" }
        },
        "required": ["id", "wordId", "word", "meaning", "planType", "createdAt"]
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 0 },
          "message": { "type": "string", "example": "ok" },
          "data": { "type": "object", "nullable": true },
          "traceId": { "type": "string", "example": "abc123xyz" }
        },
        "required": ["code", "message"]
      },
...
      "ApiResponseWordListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponseVoid" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/WordListResponse" }
            }
          }
        ]
      },
      "ApiResponseWordResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponseVoid" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/WordResponse" }
            }
          }
        ]
      },
...
      "ApiResponseStatsResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponseVoid" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/StatsResponse" }
            }
          }
        ]
      },
      "ApiResponseStudyPlanResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponseVoid" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/StudyPlanResponse" }
            }
          }
        ]
      },
      "ApiResponseListStudyPlanResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponseVoid" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/StudyPlanResponse" }
              }
            }
          }
        ]
      },
...
    }
  }
}